Assignment 2 : Complete (with bonus task) Task 1 : Address Space : Virtual Memory Support : Complete Task 2 : System Call: Exec and Exit : Complete Task 3 : System Call: Read and Write : Complete Task 4 : User Exception Handling (Bonus) : Complete I was too lazy to complete the "extra bonus task" (who cares, anyway?).
Almost all the work has been done in the "userprog" directory, with just a single modification in "threads/thread.h" (added field : threadID).

Circular dependency was a real pain in the ass to deal with during this assignment.

Note: After making any change, execute "rebuild". To run a program, execute "run program_name". Both are located in the "code" directory.